# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Groupick API
  description: |
    공급자 · 본사 · 오프라인 매장 · 고객, 4자 이해관계자 구조의 공동구매 이커머스 서비스 API입니다.

    ## 공통 응답 구조

    **성공**
    ```json
    {
      "code": "SUCCESS_CODE",
      "data": { ... }
    }
    ```

    **실패**
    ```json
    {
      "code": "ERROR_CODE",
      "errors": [
        {
          "field": "fieldName",
          "reason": "에러 사유"
        }
      ]
    }
    ```
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: 로컬 개발 서버
...
paths:
...
  /payments/prepare:
    post:
      tags: [Payment]
      summary: 결제 준비 (2-phase 1단계)
      description: |
        2-phase 결제 흐름의 1단계. 서버가 paymentId(UUID)를 생성하고 Payment PENDING 레코드를
        선(先) 저장한다. 클라이언트는 반환된 paymentId와 amount를 포트원 SDK에 전달하여 실결제를 진행한다.
      operationId: preparePayment
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreparePaymentRequest'
      responses:
        '200':
          description: 결제 준비 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreparePaymentResponse'
              example:
                code: SUCCESS
                data:
                  paymentId: "pay-uuid-001"
                  amount: 50000
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: 존재하지 않는 주문
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 소유권 불일치 또는 결제 불가 상태
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /payments/verify:
    post:
      tags: [Payment]
      summary: 결제 검증 및 확정 (2-phase 2단계, 클라이언트 호출)
      description: |
        2-phase 결제 흐름의 2단계. 포트원 SDK 결제 완료 후 클라이언트가 호출한다.
        서버 측 금액 검증 및 주문 확정을 처리하며, 이미 PAID 상태이면 멱등 응답을 반환한다.
      operationId: verifyPayment
      security:
        - BearerAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          description: 이중 결제 방지용 멱등성 키 (UUID, 추후 지원 예정)
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyPaymentRequest'
      responses:
        '200':
          description: 결제 검증 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyPaymentResponse'
              example:
                code: SUCCESS
                data:
                  orderId: 1
                  paymentId: "pay-uuid-001"
                  amount: 50000
                  status: PAID
                  paidAt: '2026-03-30T10:05:00'
                  orderStatus: PAID
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: 존재하지 않는 결제
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 결제 금액 불일치
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: PAYMENT_AMOUNT_MISMATCH
                errors:
                  - field: ~
                    reason: 결제 금액이 주문 금액과 일치하지 않습니다.
        '502':
          description: PG사 통신 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /payments/webhook:
    post:
      tags: [Payment]
      summary: PortOne 결제 웹훅 수신 (2-phase 2단계, PortOne 서버 호출)
      description: |
        PortOne 서버가 결제 완료 시 호출하는 웹훅 엔드포인트.
        PortOne V2는 Standard Webhooks 규격의 서명 헤더(webhook-id / webhook-timestamp / webhook-signature)를
        함께 전송하며, 엔드포인트는 이 서명을 검증한다. 서명이 인증을 대체하므로 유효한 서명이 없는 요청은 거부된다.
        /payments/verify와 동일한 completePayment 서비스를 호출하며 멱등 처리된다.
      operationId: receivePaymentWebhook
      parameters:
        - name: webhook-id
          in: header
          required: true
          schema:
            type: string
          description: Standard Webhooks 메시지 식별자
        - name: webhook-timestamp
          in: header
          required: true
          schema:
            type: string
          description: 웹훅 전송 시각(Unix epoch 초). ±300초 허용
        - name: webhook-signature
          in: header
          required: true
          schema:
            type: string
          description: 공백 구분 "{버전},{base64 HMAC-SHA256 서명}" 토큰 목록
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PortOneWebhookPayload'
            example:
              type: Transaction.Paid
              data:
                paymentId: "pay-uuid-001"
      responses:
        '200':
          description: 웹훅 처리 완료
        '400':
          description: 웹훅 서명 검증 실패 (PAYMENT_010)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: WEBHOOK_VERIFICATION_FAILED
                errors:
                  - field: ~
                    reason: 웹훅 검증에 실패했습니다

  # ──────────────────────────────────────────
  # Notification
  # ──────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 로그인 후 발급된 JWT access_token을 입력하세요.
...
  responses:
    Unauthorized:
      description: 인증 토큰이 없거나 만료되었습니다.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: AUTH_UNAUTHORIZED
            errors: []
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          description: 에러 코드
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              reason:
                type: string
...
    PreparePaymentRequest:
      type: object
      required: [order_id]
      properties:
        order_id:
          type: integer
          format: int64
    PreparePaymentResponse:
      type: object
      properties:
        code:
          type: string
        data:
          type: object
          properties:
            paymentId:
              type: string
              description: 서버 생성 UUID (포트원 SDK에 전달)
            amount:
              type: integer
              format: int64
    VerifyPaymentRequest:
      type: object
      required: [order_id, payment_id]
      properties:
        order_id:
          type: integer
          format: int64
        payment_id:
          type: string
          description: preparePayment에서 반환된 paymentId
    VerifyPaymentResponse:
      type: object
      properties:
        code:
          type: string
        data:
          type: object
          properties:
            orderId:
              type: integer
              format: int64
            paymentId:
              type: string
            amount:
              type: integer
              format: int64
            status:
              type: string
              enum: [PAID, CANCELLED]
            paidAt:
              type: string
              format: date-time
            orderStatus:
              type: string
              enum: [PAID, CANCELLED]
    PortOneWebhookPayload:
      type: object
      required: [type, data]
      properties:
        type:
          type: string
          example: Transaction.Paid
        data:
          type: object
          required: [paymentId]
          properties:
            paymentId:
              type: string

    # ── Notification ──────────────────────
...